requirements.txt
Flask 
Flask-SQLAlchemy
pymysql
psycopg2
cryptography
flask-migrate
pandas
requests
python-dotenv
flask_cors
flask-bcrypt
pyjwt
pillow
Flask-Mail